+++ src/util/index.ts
@@ -1,6 +1,7 @@
 import { BigNumber, BigNumberish } from "@ethersproject/bignumber";
 import { AddressUtil } from "./address";
 import { CoinUtil } from "./coin";
+import { GobyUtil } from "./goby";
 import { NetworkUtil } from "./network";
 import { SerializerUtil } from "./serializer";
 import { SExpUtil } from "./sexp";
@@ -11,6 +12,7 @@ export class Util {
     public static serializer: SerializerUtil = new SerializerUtil();
     public static network: NetworkUtil = new NetworkUtil();
     public static sexp: SExpUtil = new SExpUtil();
+    public static goby: GobyUtil = new GobyUtil();
     public static mojoPerXCH: BigNumber = BigNumber.from(1000000000000);
 
     public static formatToken(amount: BigNumberish, amountPerUnit: BigNumberish = 1000): string {