+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
 ## 1.1.6
+ - Modify return types of the `Provider.tranfer`, `Provider.transferCAT`, and `Provider.acceptOffer` methods from `Promise<boolean>` to `Promise<Optional<SpendBundle>>`
  - Add `greenweb.util.sexp.curry()`
  - Add `SmartCoin.curry()`
+ - Create `greenweb.util.goby`